+++ Examples/Examples/ESTCloudBeaconTableVC.m
@@ -103,7 +103,7 @@
     cell.textLabel.text         = [NSString stringWithFormat:@"Major: %@, Minor: %@", beacon.major, beacon.minor];
     cell.detailTextLabel.text   = [NSString stringWithFormat:@"UUID: %@", beacon.UUID ];
     
-    cell.imageView.image = [UIImage imageNamed:@"beacon"];
+    cell.imageView.image = beacon.isSecured ? [UIImage imageNamed:@"beacon_secure"] : [UIImage imageNamed:@"beacon"];
     
     return cell;
 }